simplify and cleanup the block for turning on AF instr; share a test utility function
