+++ lib/instrumentation/index.js
@@ -228,20 +228,9 @@ Instrumentation.prototype.start = function (runContextClass) {
     undiciInstr.instrumentUndici(this._agent)
   }
 
-  if (
-    azureFunctionsInstr.isAzureFunctionsEnvironment &&
-    // We are assuming FUNCTIONS_EXTENSION_VERSION=~4.
-    // https://aka.ms/functions-node-versions
-    // XXX How to watch/maintain this? Can bitrot.js watch for this? Is there harm in just always turning this on? E.g. try with node v20.
-    semver.satisfies(process.version, '>=14 <19')
-  ) {
+  if (azureFunctionsInstr.isAzureFunctionsEnvironment) {
     this._log.debug('instrumenting azure-functions')
-    // setInterval(() => {
-    //   this._log.flush() // XXX not sure this is doing anything
-    // }, 10 * 1000)
     azureFunctionsInstr.instrument(this._agent)
-  } else {
-    this._log.info('XXX NOT instr azure-functions', this._isModuleEnabled('azure-functions'), process.env.FUNCTIONS_WORKER_RUNTIME, semver.satisfies(process.version, '>=14 <19'))
   }
 }
