+++ test/instrumentation/modules/next/next.test.js
@@ -31,6 +31,7 @@ const semver = require('semver')
 const tape = require('tape')
 
 const { MockAPMServer } = require('../../../_mock_apm_server')
+const { formatForTComment } = require('../../_utils')
 
 if (os.platform() === 'win32') {
   // Limitation: currently don't support testing on Windows.
@@ -55,9 +56,6 @@ if (process.env.ELASTIC_APM_CONTEXT_MANAGER === 'patch') {
 }
 
 const testAppDir = path.join(__dirname, 'a-nextjs-app')
-
-// Match ANSI escapes (from https://stackoverflow.com/a/29497680/14444044).
-const ANSI_RE = /[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g /* eslint-disable-line no-control-regex */
 
 let apmServer
 let nextJsVersion // Determined after `npm ci` is run.
@@ -390,23 +388,6 @@ if (DEV_TEST_FILTER) {
 }
 
 // ---- utility functions
-
-/**
- * Format the given data for passing to `t.comment()`.
- *
- * - t.comment() wipes leading whitespace. Prefix lines with '|' to avoid
- *   that, and to visually group a multi-line write.
- * - Drop ANSI escape characters, because those include control chars that
- *   are illegal in XML. When we convert TAP output to JUnit XML for
- *   Jenkins, then Jenkins complains about invalid XML. `FORCE_COLOR=0`
- *   can be used to disable ANSI escapes in `next dev`'s usage of chalk,
- *   but not in its coloured exception output.
- */
-function formatForTComment (data) {
-  return data.toString('utf8')
-    .replace(ANSI_RE, '')
-    .trimRight().replace(/\n/g, '\n|') + '\n'
-}
 
 /**
  * Wait for the test a-nextjs-app server to be ready.
